added font asm

+++ logotable.h
@@ -6,7 +6,7 @@
 #include "util.h"
 
 void initLogotable();
-uint8 getNextLogoRunlen();
+uint16 getNextLogoRunlen();
 
 
 #endif	/* LOGOTABLE_H */
